updated to new namespace naming convention and file naming convention
