+++ .Rhistory
@@ -1,4 +1,3 @@
-.Call(cmdline.arguments:::argument_locator_test, x, y)
 library(cmdline.arguments)
 .Call(cmdline.arguments:::argument_locator_test, x, y)
 x
@@ -510,3 +509,4 @@ sum(s1 * s2 * s3 %*% s4 * s5)
 s3
 s3 %*% s4
 sum(s1 * s2 * (s3 %*% s4) * s5)
+?e1071::svm
