+++ config.log
@@ -60,6 +60,7 @@ PATH: /c/Users/olive/AppData/Local/Microsoft/WindowsApps
 PATH: /c/Users/olive/AppData/Local/Programs/Microsoft VS Code Insiders/bin
 PATH: /c/Users/olive/.dotnet/tools
 PATH: /c/Users/olive/AppData/Roaming/TinyTeX/bin/win32
+PATH: /c/Program Files/RStudio/bin/postback
 PATH: /bin
 PATH: /c/Program Files/RStudio/bin/msys-ssh-1000-18
 
@@ -85,7 +86,7 @@ generated by GNU Autoconf 2.69.  Invocation command line was
 
 on LAPTOP-H6P1GJFG
 
-config.status:713: creating src/Makevars
+config.status:714: creating src/Makevars
 
 ## ---------------- ##
 ## Cache variables. ##
@@ -117,7 +118,7 @@ PACKAGE_URL=''
 PACKAGE_VERSION=''
 PATH_SEPARATOR=':'
 SHELL='/bin/sh'
-SUBDIR_SOURCES='exports.cpp parser/argument/argument.cpp utils/cmd_coerce.cpp utils/exceptions.cpp utils/matching.cpp utils/Rapi.cpp '
+SUBDIR_SOURCES='exports.cpp parser/argument/class.cpp parser/argument/raw/class.cpp parser/narg/narg.cpp parser/pfunc/pfunc.cpp utils/converters/converters.cpp utils/Rapi.cpp utils/type_name/type_name.cpp '
 bindir='${exec_prefix}/bin'
 build_alias=''
 datadir='${datarootdir}'
