+++ src/parser/narg/narg.h
@@ -20,67 +20,61 @@ using Rcpp::stop;
  */
 
 
-namespace cmdline_arguments::parser{
+namespace cmd_args{
+  namespace parser{
+    class narg{
+    private:
+      // storage for raw "narg" and name (latter used for error message)
+      // TODO: change name to a string reference, once name is stored in
+      // an overall container safely (breaks when not stored safely)
+      const string narg_raw;
+      const string name; // Should work as we don't
+      // store value.
+      R_xlen_t current_value = 0, critical_value;
 
-  class narg{
-  private:
-    // storage for raw "narg" and name (latter used for error message)
-    // TODO: change name to a string reference, once name is stored in
-    // an overall container safely (breaks when not stored safely)
-    const string narg_raw;
-    const string name; // Should work as we don't
-    // store value.
-    R_xlen_t current_value = 0, critical_value;
-
-    // Comparison functions. Store the one we'll actually use in the function
-    // to minimize time spent in "if" and "switch". (we only need 1 dereferencing here)
-    //
-    //typedef bool (narg::*cmpfun)(R_xlen_t& more) const;
-    //cmpfun compare_fun;
+      // Comparison functions. Store the one we'll actually use in the function
+      // to minimize time spent in "if" and "switch". (we only need 1 dereferencing here)
+      std::function<bool(R_xlen_t&)> compare_fun;
 
 
-    // I would prefer to do as such:But for some reason this fails.
-    std::function<bool(R_xlen_t&)> compare_fun;
+      //function<bool(R_xlen_t&)> compare_fun;
+      bool le(R_xlen_t&) const,
+        leq(R_xlen_t&) const,
+        ge(R_xlen_t&) const,
+        geq(R_xlen_t&) const,
+        eq(R_xlen_t&) const;
+      // Error message (set in parse_narg):
+      string errormsg;
+      void throwerr() const;
 
+      // function for parsing input arguments.
+      void parse_narg();
 
-    //function<bool(R_xlen_t&)> compare_fun;
-    bool le(R_xlen_t&) const,
-      leq(R_xlen_t&) const,
-      ge(R_xlen_t&) const,
-      geq(R_xlen_t&) const,
-      eq(R_xlen_t&) const;
-    // Error message (set in parse_narg):
-    string errormsg;
-    void throwerr() const;
+    public:
+      // Initializer
+      narg(const string, const string);
+      // comparison function
+      // Test if we can accept 'R_xlen_t' more arguments.
+      bool digest(R_xlen_t) const;
+      // Close narg,
+      // Test that the bool is fulfilled, otherwise throw error.
+      void close() const;
 
-    // function for parsing input arguments.
-    void parse_narg();
+      void add(R_xlen_t);
 
-  public:
-    // Initializer
-    narg(const string, const string);
-    // comparison function
-    // Test if we can accept 'R_xlen_t' more arguments.
-    bool digest(R_xlen_t) const;
-    // Close narg,
-    // Test that the bool is fulfilled, otherwise throw error.
-    void close() const;
+      // Operator overload to "digest"
+      bool operator()(R_xlen_t&) const;
+      bool operator()() const;
+      // Another way of calling "add".
+      void operator+=(R_xlen_t&);
+      void operator++();
 
-    void add(R_xlen_t);
-
-    // Operator overload to "digest"
-    bool operator()(R_xlen_t&) const;
-    bool operator()() const;
-    // Another way of calling "add".
-    void operator+=(R_xlen_t&);
-    void operator++();
-
-    // getters
-    string getnarg(),
-      geterr(),
-      getname();;
-    R_xlen_t getv(), getc();
-  };
-
-}
+      // getters
+      string getnarg(),
+        geterr(),
+        getname();;
+      R_xlen_t getv(), getc();
+    };
+  } // ~ parser
+} // ~ cmd_args
 #endif
