+++ src/utils/Rapi.cpp
@@ -1,9 +1,13 @@
 
 #include "utils/Rapi.h"
 
-namespace cmdline_arguments::utils::Rapi{
-  //template <typename T>
-  bool cmd_ISNULL(SEXP sexp){
-    return sexp == R_NilValue || TYPEOF(sexp) == NILSXP;
+namespace cmd_args::utils::Rapi{
+  namespace utils{
+    namespace Rapi{
+      //template <typename T>
+      bool cmd_ISNULL(SEXP sexp){
+        return sexp == R_NilValue || TYPEOF(sexp) == NILSXP;
+      }
+    }
   }
 }
