+++ src/utils/Rapi.h
@@ -2,8 +2,12 @@
 #ifndef cmd_Rapi
 #define cmd_Rapi
 #include <Rinternals.h>
-namespace cmdline_arguments::utils::Rapi{
-  //template <typename T>
-  bool cmd_ISNULL(SEXP);
+namespace cmd_args::utils::Rapi{
+  namespace utils{
+    namespace Rapi{
+      //template <typename T>
+      bool cmd_ISNULL(SEXP);
+    }
+  }
 }
 #endif
